+++ waspc/cli/src/Wasp/Cli/Command/Db/Seed.hs
@@ -19,8 +19,10 @@ import Wasp.Cli.Command (Command, CommandError (CommandError), require)
 import Wasp.Cli.Command.Compile (analyze)
 import Wasp.Cli.Command.Message (cliSendMessageC)
 import Wasp.Cli.Command.Require.InWaspProject (InWaspProject (InWaspProject))
+import Wasp.Cli.Util.Apps (defaultAppPorts, getWaspEnvVars)
 import Wasp.Generator.DbGenerator.Operations (dbSeed)
 import qualified Wasp.Message as Msg
+import Wasp.Project.Apps (server)
 import Wasp.Project.Common (generatedAppDirInWaspProjectDir)
 
 seed :: Maybe String -> Command ()
@@ -34,7 +36,9 @@ seed maybeUserProvidedSeedName = do
 
   cliSendMessageC $ Msg.Start $ "Running database seed " <> nameOfSeedToRun <> "..."
 
-  liftIO (dbSeed genProjectDir nameOfSeedToRun) >>= \case
+  let waspEnvVars = server $ getWaspEnvVars appSpec defaultAppPorts
+
+  liftIO (dbSeed waspEnvVars genProjectDir nameOfSeedToRun) >>= \case
     Left errorMsg -> E.throwError $ CommandError "Database seeding failed" errorMsg
     Right () -> cliSendMessageC $ Msg.Success "Database seeded successfully!"